fix: rewrite of list item normalizer
Instead of checking in child of list whether it's an element, we now
check that list has all children as elements.
This prevents unintended wrapping of multiple text nodes (as would be
the case for different marks on the same line) in same number of
paragraphs. Original caused a bug, where single list item with
multiple marks was spread onto multiple lines, one per mark.